requirements.txt
pandas==2.0.3
numpy==1.24.3
google-cloud-aiplatform==1.35.0
google-cloud-storage==2.13.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
streamlit==1.28.0
kfp==2.3.0
protobuf==4.24.4
